moved grouping strategy specs from GroupServiceSpec to GroupingStrategySpec
